+++ src/test/scala/cognite/spark/v1/SdkV1RddTest.scala
@@ -37,7 +37,12 @@ class SdkV1RddTest extends FlatSpec with Matchers with ParallelTestExecution wit
     def uniqueId(s: String): String = "1"
 
     val sdkRdd =
-      SdkV1Rdd(spark.sparkContext, getDefaultConfig(CdfSparkAuth.Static(readApiKeyAuth)), toRow, uniqueId, getStreams)
+      SdkV1Rdd(
+        spark.sparkContext,
+        getDefaultConfig(CdfSparkAuth.Static(readApiKeyAuth)),
+        toRow,
+        uniqueId,
+        getStreams)
 
     val e = intercept[CdpApiException] {
       sdkRdd.compute(CdfPartition(0), TaskContext.get())
@@ -47,21 +52,19 @@ class SdkV1RddTest extends FlatSpec with Matchers with ParallelTestExecution wit
   }
 
   it should "convert multiple streams to one Iterator" in {
-    import CdpConnector._
-
     val nStreams = 50
     val nItemsPerStream = 1000
     val rdd = new SdkV1Rdd[Event, Long](
       spark.sparkContext,
       DefaultSource
         .parseRelationConfig(Map("apiKey" -> writeApiKey), spark.sqlContext)
-        .copy(parallelismPerPartition = nStreams),
+        .copy(parallelismPerPartition = nStreams * 3),
       (e: Event, partitionIndex: Option[Int]) => asRow(e),
       (e: Event) => e.id,
       (_: GenericClient[IO], _: Option[Int], _: Int) => {
         val allStreams = 0.until(nStreams).map { i =>
           Stream.evalUnChunk {
-            IO.sleep((scala.math.random * 300).millis)(cdpConnectorTimer) *> IO(
+            IO.sleep((scala.math.random * 300).millis) *> IO(
               Chunk.seq(1.to(nItemsPerStream).map(j => Event(id = i * nItemsPerStream + j))))
           }
         }
@@ -70,6 +73,35 @@ class SdkV1RddTest extends FlatSpec with Matchers with ParallelTestExecution wit
       }
     )
 
+    assert(rdd.partitions.length == 1)
     assert(rdd.compute(CdfPartition(0), TaskContext.get()).size == nStreams * nItemsPerStream)
+  }
+
+  it should "convert multiple streams and keep duplicate" in {
+    import CdpConnector._
+
+    val nStreams = 50
+    val nItemsPerStream = 1000
+    val rdd = new SdkV1Rdd[Event, Long](
+      spark.sparkContext,
+      DefaultSource
+        .parseRelationConfig(Map("apiKey" -> writeApiKey), spark.sqlContext)
+        .copy(parallelismPerPartition = nStreams * 3),
+      (e: Event, partitionIndex: Option[Int]) => asRow(e),
+      (e: Event) => e.id,
+      (_: GenericClient[IO], _: Option[Int], _: Int) => {
+        val allStreams = 0.until(nStreams).map { i =>
+          Stream.evalUnChunk {
+            IO.sleep((scala.math.random * 300).millis) *> IO(
+              Chunk.seq(1.to(nItemsPerStream).map(j => Event(id = i * nItemsPerStream + j))))
+          }
+        }
+        allStreams ++ allStreams ++ allStreams
+      },
+      deduplicateRows = false
+    )
+
+    assert(rdd.partitions.length == 1)
+    assert(rdd.compute(CdfPartition(0), TaskContext.get()).size == nStreams * nItemsPerStream * 3) // * 3 because we duplicate the allStreams 3 times
   }
 }